api/requirements.txt
Django==3.1.*
djangorestframework
markdown
django-filter
django-guardian
Pygments
python-dotenv
gunicorn
autopep8
django-sass
Pillow
django-cors-headers